add default config for evo70
