+++ keyboards/custommk/evo70/config.h
@@ -19,17 +19,6 @@
 #include "config_common.h"
 
 #define OLED_UPDATE_INTERVAL 33
-
-#define BACKLIGHT_PIN B5
-// #define BACKLIGHT_LIMIT_VAL 255
-#define BACKLIGHT_ON_STATE 0
-#define BACKLIGHT_LEVELS 5
-#define BACKLIGHT_DEFAULT_LEVEL 0
-#define BACKLIGHT_CAPS_LOCK
-
-// #define BACKLIGHT_BREATHING
-// #define BREATHING_PERIOD 6
-
 #define OLED_DISABLE_TIMEOUT
 
 #define ENCODERS_PAD_A { C7 }
@@ -37,20 +26,25 @@
 #define ENCODER_RESOLUTION 2
 #define TAP_CODE_DELAY 10
 
+#define BACKLIGHT_PIN B5
+#define BACKLIGHT_ON_STATE 0
+#define BACKLIGHT_LEVELS 5
+#define BACKLIGHT_DEFAULT_LEVEL 3
+#define BACKLIGHT_CAPS_LOCK
+
 // #define RGBLIGHT_EFFECT_BREATHING
 // #define RGBLIGHT_EFFECT_RAINBOW_MOOD
-#define RGBLIGHT_EFFECT_RAINBOW_SWIRL
+// #define RGBLIGHT_EFFECT_RAINBOW_SWIRL
 // #define RGBLIGHT_EFFECT_SNAKE
-#define RGBLIGHT_EFFECT_KNIGHT
+// #define RGBLIGHT_EFFECT_KNIGHT
 // #define RGBLIGHT_EFFECT_CHRISTMAS
-// #define RGBLIGHT_EFFECT_STATIC_GRADIENT
+#define RGBLIGHT_EFFECT_STATIC_GRADIENT
 // #define RGBLIGHT_EFFECT_RGB_TEST
 // #define RGBLIGHT_EFFECT_ALTERNATING
-// #define RGBLIGHT_EFFECT_TWINKLE // 648
+// #define RGBLIGHT_EFFECT_TWINKLE
 
 // #define RGBLIGHT_LAYERS
 #define RGBLIGHT_VAL_STEP 25
 #define RGBLIGHT_LIMIT_VAL 250
 #define RGBLIGHT_DEFAULT_VAL 150
-#define RGBLIGHT_DEFAULT_MODE RGBLIGHT_EFFECT_RAINBOW_SWIRL + 4
 #define RGBLIGHT_SLEEP
