+++ keyboards/custommk/evo70/keymaps/passing-comet/readme.md
@@ -1,0 +1,3 @@
+# EVO70 Layout for Vial
+
+This is the Vial layout for the EVO70. It is identical to the default layout except Vial is enabled.
