+++ keyboards/custommk/evo70/keymaps/passing-comet/rules.mk
@@ -1,9 +1,12 @@
 VIA_ENABLE = yes
 VIAL_ENABLE = yes
 VIAL_INSECURE = yes
-
 QMK_SETTINGS = no # 2486
-TAP_DANCE_ENABLE = no # 1812
+
+# BACKLIGHT_ENABLE = no
+SPACE_CADET_ENABLE = no # ~324
+AUTO_SHIFT_ENABLE = no # 2528
+
 COMBO_ENABLE = no # ~2408
+TAP_DANCE_ENABLE = yes # 1812
 KEY_OVERRIDE_ENABLE = no # 1846
-BACKLIGHT_ENABLE = no
