+++ keyboards/custommk/evo70/keymaps/vial/config.h
@@ -1,14 +1,12 @@
-#define VIAL_KEYBOARD_UID {0x6E, 0x30, 0x6F, 0x9F, 0xE0, 0xBF, 0xAC, 0x6D}
-#define VIAL_TAP_DANCE_ENTRIES 16
-// #define VIAL_KEY_OVERRIDE_ENTRIES 8
+/* SPDX-License-Identifier: GPL-2.0-or-later */
 
-// #define DYNAMIC_KEYMAP_LAYER_COUNT 3
-#define DEBOUNCE 7
-#define TAPPING_TERM 175
+#pragma once
+
+#define VIAL_KEYBOARD_UID {0x6E, 0x30, 0x6F, 0x9F, 0xE0, 0xBF, 0xAC, 0x6D}
+
+#define DEBOUNCE 5
+#define TAPPING_TERM 200
 
 #define NO_ACTION_ONESHOT
 
-// #define NO_DEBUG
-// #define NO_PRINT
-// #define VIAL_UNLOCK_COMBO_ROWS { 0, 2 }
-// #define VIAL_UNLOCK_COMBO_COLS { 0, 13 }
+#define RGBLIGHT_DEFAULT_MODE RGBLIGHT_EFFECT_STATIC_GRADIENT + 0
