+++ keyboards/custommk/evo70/keymaps/vial/readme.md
@@ -1,3 +1,3 @@
 # EVO70 Layout for Vial
 
-This is the Vial layout for the EVO70. It is identical to the default layout except Vial is enabled.
+This is the Vial layout for the EVO70.
