+++ keyboards/custommk/evo70/keymaps/vial/rules.mk
@@ -1,12 +1,9 @@
 VIA_ENABLE = yes
 VIAL_ENABLE = yes
+
 VIAL_INSECURE = yes
-QMK_SETTINGS = no # 2486
+QMK_SETTINGS = yes
 
-# BACKLIGHT_ENABLE = no
-SPACE_CADET_ENABLE = no # ~324
-AUTO_SHIFT_ENABLE = no # 2528
-
-COMBO_ENABLE = no # ~2408
-TAP_DANCE_ENABLE = yes # 1812
-KEY_OVERRIDE_ENABLE = no # 1846
+COMBO_ENABLE = no
+TAP_DANCE_ENABLE = no
+KEY_OVERRIDE_ENABLE = no
