--- keyboards/custommk/evo70/keymaps/vial2/config.h
@@ -1,3 +1,0 @@
-#define VIAL_KEYBOARD_UID {0x6E, 0x30, 0x6F, 0x9F, 0xE0, 0xBF, 0xAC, 0x6D}
-#define DEBOUNCE 6
-#define NO_ACTION_ONESHOT
