--- keyboards/custommk/evo70/keymaps/vial2/readme.md
@@ -1,3 +1,0 @@
-# EVO70 Layout for VIA
-
-This is the via layout for the EVO70. It is identical to the default layout except VIA is enabled.
